server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
pydantic==2.5.3
python-dotenv==1.0.0
